Moving plotting tools into utils
